+++ test/runtests.jl
@@ -6,6 +6,7 @@ using LinearAlgebra
     include("test_aqua.jl")
     include("test_readme.jl")
     include("test_golden.jl")
+    include("test_amplification.jl")
     include("test_convergence.jl")
     include("test_invariants.jl")
     include("test_symmetry.jl")